requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
pyodbc==4.0.35
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
